Split docker context initialization
